+++ src/OpenTelemetry.Exporter.Jaeger/JaegerExporterHelperExtensions.cs
@@ -62,10 +62,15 @@ namespace OpenTelemetry.Trace
 
             name ??= Options.DefaultName;
 
-            if (configure != null)
+            builder.ConfigureServices(services =>
             {
-                builder.ConfigureServices(services => services.Configure(name, configure));
-            }
+                if (configure != null)
+                {
+                    services.Configure(name, configure);
+                }
+
+                services.AddSingleton<IOptionsFactory<JaegerExporterOptions>, JaegerExporterOptions.JaegerExporterOptionsFactory>();
+            });
 
             return builder.ConfigureBuilder((sp, builder) =>
             {